vqa_app/backend/requirements.txt
fastapi
peft
uvicorn[standard]
torch
transformers==4.46.0
Pillow
python-multipart